+++ CHANGELOG.md
@@ -12,3 +12,15 @@
 . **noctrl**, to remove all MIDI CC from each and every instrument
 
 __
+
+## 1.2.1
+
+. Validations in some helper functions
+
+. Improved sync
+
+. New helper functions - **edger**, **edgerx**, **swarm**, **chop**
+
+. Instruments +1 to work as a metronome, alias **sync**
+
+. Rename **istep** to **isync**